+++ examples/refund.php
@@ -18,6 +18,7 @@ declare(strict_types=1);
 require __DIR__ . '/../src/P2FluxException.php';
 require __DIR__ . '/../src/ChargeResult.php';
 require __DIR__ . '/../src/P2FluxClient.php';
+require __DIR__ . '/../src/CurlTransport.php';
 
 use P2Flux\P2FluxClient;